requirements.txt
autogen
yfinance
matplotlib
flask
google-generativeai